requirements.txt
astroid==2.4.0
attrs==19.3.0
autopep8==1.5.2
certifi==2020.4.5.1
importlib-metadata==1.6.0
isort==4.3.21
kafka-python==2.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pycodestyle==2.5.0
pylint==2.5.0
pyparsing==2.4.7
pytest==5.4.1
python-snappy==0.5.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
